+++ apps/wallop_web/lib/wallop_web/application.ex
@@ -5,7 +5,6 @@ defmodule WallopWeb.Application do
   @impl true
   def start(_type, _args) do
     children = [
-      {Phoenix.PubSub, name: WallopWeb.PubSub},
       WallopWeb.Endpoint
     ]